+++ src/main/resources/lessons/authbypass/documentation/lesson-template-video.adoc
