+++ src/main/resources/lessons/authbypass/images/firefox-proxy-config.png
