+++ src/main/resources/lessons/authbypass/images/paypal-2fa-bypass.png
