+++ src/main/resources/lessons/authbypass/js/bypass.js
