+++ src/test/java/org/owasp/webgoat/lessons/authbypass/BypassVerificationTest.java
@@ -23,13 +23,14 @@
  * <p>
  */
 
-package org.owasp.webgoat.lessons.auth_bypass;
+package org.owasp.webgoat.lessons.authbypass;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.owasp.webgoat.container.assignments.AssignmentEndpointTest;
+import org.owasp.webgoat.lessons.authbypass.VerifyAccount;
 import org.springframework.test.web.servlet.MockMvc;
 
 import static org.springframework.test.web.servlet.setup.MockMvcBuilders.standaloneSetup;
